+++ api.c
@@ -33,6 +33,7 @@
 #include "libtcmu_log.h"
 #include "libtcmu_common.h"
 #include "libtcmu_priv.h"
+#include "target.h"
 #include "alua.h"
 
 int tcmu_get_cdb_length(uint8_t *cdb)
